Update selinux-policy to version 20230728 / rev 49 via SR 1115652
https://build.opensuse.org/request/show/1115652
by user cahu + anag+factory

+++ packages/s/selinux-policy/update.sh
@@ -2,18 +2,20 @@
 
 date=$(date '+%Y%m%d')
 base_name_pattern='selinux-policy-*.tar.xz'
-
 echo Update to $date
 
 old_tar_file=$(ls -1 $base_name_pattern)
 
 osc service manualrun
 
-rm -rf container-selinux
-git clone --depth 1 https://github.com/containers/container-selinux.git
-rm -f container.*
-mv container-selinux/container.* .
-rm -rf container-selinux
+if [ "$1" = "full" ]; then
+  echo doing full update including container-selinux
+  rm -rf container-selinux
+  git clone --depth 1 https://github.com/containers/container-selinux.git
+  rm -f container.*
+  mv container-selinux/container.* .
+  rm -rf container-selinux
+fi
 
 # delete old files. Might need a better sanity check
 tar_cnt=$(ls -1 $base_name_pattern  | wc -l)
@@ -24,4 +26,3 @@ if [ $tar_cnt -gt 1 ]; then
 fi
 
 osc status
-
